moved MscGroupBox in ui.widgets.msc_groupbox.py
